Refactored remap of kernel into memory::init
